+++ nodes.py
@@ -2936,6 +2936,7 @@ class AudioLoopHelperExtension(ComfyExtension):
     async def get_node_list(self) -> list[type[io.ComfyNode]]:
         from .nodes_analysis import AudioPitchDetect
         from .nodes_easycache import LTXVideoEasyCache
+        from .nodes_regional_compile import LTXVideoRegionalCompile
         from .nodes_sage import AudioLoopHelperSageAttention
         from .nodes_validation import LoopConfigValidator
 
@@ -2964,6 +2965,7 @@ class AudioLoopHelperExtension(ComfyExtension):
             LoopIterationStamp,
             AudioLoopHelperSageAttention,
             LTXVideoEasyCache,
+            LTXVideoRegionalCompile,
             ProfileBegin,
             ProfileIterStep,
             ProfileEnd,